module: cfi: ensure __cfi_check alignment
On modules with no executable code, LLVM generates a __cfi_check stub,
but won't align it to page size as expected. This change ensures the
function is at the beginning of the .text section and correctly aligned
for the CFI shadow.

Also fixing LD_FLAGS_LTO_CLANG being passed 2 times to LLVM LD

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>